daemon/requirements.txt
python-telegram-bot[ext]>=21.0
openai>=1.0.0
aiosqlite>=0.19.0
pyyaml>=6.0
aiohttp>=3.9.0
apscheduler>=3.10.0
python-dotenv>=1.0.0
textual>=0.85.0
playwright>=1.40.0
pyjwt>=2.8.0
cryptography>=42.0.0
httpx>=0.27.0
